refactor: considerable amount of refactor, clean up test structure
