+++ .aegir.js
@@ -1,61 +1,65 @@
 'use strict'
 
-const Node = require('./test/nodejs-bundle/nodejs-bundle.js')
 const PeerInfo = require('peer-info')
 const PeerId = require('peer-id')
 const pull = require('pull-stream')
-const {parallel} = require('async')
+const parallel = require('async/parallel')
 
+const rawPeer = require('./test/fixtures/test-peer.json')
+const Node = require('./test/utils/bundle.node.js')
 const sigServer = require('libp2p-webrtc-star/src/sig-server')
-const wsRendezvous = require('libp2p-websocket-star-rendezvous')
-let server
-let server2
+const WebSocketStarRendezvous = require('libp2p-websocket-star-rendezvous')
 
+let wrtcRendezvous
+let wsRendezvous
 let node
-const rawPeer = require('./test/browser-bundle/peer.json')
 
 const before = (done) => {
-  let count = 0
-  const ready = () => ++count === 2 ? done() : null
+  parallel([
+    (cb) => {
+      sigServer.start({ port: 15555 }, (err, server) => {
+        if (err) { return cb(err) }
+        wrtcRendezvous = server
+        cb()
+      })
+    },
+    (cb) => {
+      WebSocketStarRendezvous.start({
+        port: 14444,
+        refreshPeerListIntervalMS: 1000,
+        strictMultiaddr: false,
+        disableCryptoChallenge: true
+      }, (err, _server) => {
+        if (err) {
+          return cb(err)
+        }
+        wsRendezvous = _server
+        cb()
+      })
+    },
+    (cb) => {
+      PeerId.createFromJSON(rawPeer, (err, peerId) => {
+        if (err) {
+          return done(err)
+        }
+        const peer = new PeerInfo(peerId)
 
-  sigServer.start({
-    port: 15555
-  }, (err, _server) => {
-    if (err) {
-      throw err
+        peer.multiaddrs.add('/ip4/127.0.0.1/tcp/9200/ws')
+
+        node = new Node(peer)
+        node.handle('/echo/1.0.0', (protocol, conn) => pull(conn, conn))
+        node.start(cb)
+      })
     }
-    server = _server
-    ready()
-  })
-
-  wsRendezvous.start({
-    port: 14444,
-    refreshPeerListIntervalMS: 1000,
-    strictMultiaddr: false
-  }, (err, _server) => {
-    if (err) {
-      throw err
-    }
-    server2 = _server
-    ready()
-  })
-
-  PeerId.createFromJSON(rawPeer, (err, peerId) => {
-    if (err) {
-      return done(err)
-    }
-    const peer = new PeerInfo(peerId)
-
-    peer.multiaddrs.add('/ip4/127.0.0.1/tcp/9200/ws')
-
-    node = new Node(peer)
-    node.handle('/echo/1.0.0', (protocol, conn) => pull(conn, conn))
-    node.start(() => ready())
-  })
+  ], done)
 }
 
 const after = (done) => {
-  setTimeout(() => parallel([node, server, server2].map(s => cb => s.stop(cb)), done), 2000)
+  setTimeout(() => parallel(
+    [node, wrtcRendezvous, wsRendezvous].map((s) => {
+      return (cb) => s.stop(cb)
+    })
+   , done), 2000)
 }
 
 module.exports = {
