+++ test/browser.js
@@ -1,8 +1,4 @@
 'use strict'
 
-const w = require('webrtcsupport')
-
 require('./base')
-require('./browser-bundle/websockets-only')
-if (w.support) { require('./browser-bundle/webrtc-star-only') }
-require('./browser-bundle/websocket-star-only')
+require('./transports.browser')
