+++ test/node.js
@@ -1,12 +1,9 @@
 'use strict'
 
 require('./base')
-require('./nodejs-bundle/tcp')
-require('./nodejs-bundle/tcp+websockets')
-require('./nodejs-bundle/tcp+websockets+webrtc-star')
-require('./nodejs-bundle/tcp+websockets+websocket-star')
-require('./nodejs-bundle/stream-muxing')
-require('./nodejs-bundle/discovery')
-require('./nodejs-bundle/peer-routing')
-require('./nodejs-bundle/content-routing')
-require('./nodejs-bundle/circuit')
+require('./transports.node')
+require('./stream-muxing.node')
+require('./peer-discovery.node')
+require('./peer-routing.node')
+require('./content-routing.node')
+require('./circuit-relay.node')
